% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceas-package.R
\docType{package}
\name{ceas-package}
\alias{ceas}
\alias{ceas-package}
\title{ceas: Cellular Energetics Analysis Software}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Measuring cellular energetics is essential to understanding a matrix’s (e.g. cell, tissue or biofluid) metabolic state. The Agilent Seahorse machine is a common method to measure real-time cellular energetics, but existing analysis tools are highly manual or lack functionality. The Cellular Energetics Analysis Software (ceas) R package fills this analytical gap by providing modular and automated Seahorse data analysis and visualization using the methods described by Mookerjee et al. (2017) \doi{10.1074/jbc.m116.774471}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jamespeapen.github.io/ceas/}
  \item \url{https://github.com/jamespeapen/ceas/}
  \item Report bugs at \url{https://github.com/jamespeapen/ceas/issues/}
}

}
\author{
\strong{Maintainer}: Rachel House \email{rachel.house@vai.org} (\href{https://orcid.org/0000-0003-2176-0431}{ORCID})

Authors:
\itemize{
  \item James P. Eapen \email{james.eapen@vai.org} (\href{https://orcid.org/0000-0001-6016-3598}{ORCID})
}

Other contributors:
\itemize{
  \item Hui Shen \email{hui.shen@vai.org} (\href{https://orcid.org/0000-0001-9767-4084}{ORCID}) [funder]
  \item Carrie R. Graveel \email{carrie.graveel@vai.org} (\href{https://orcid.org/0000-0001-7251-5642}{ORCID}) [funder]
  \item Matthew R. Steensma \email{matthew.steensma@vai.org} (\href{https://orcid.org/0000-0002-9003-6730}{ORCID}) [funder]
  \item Van Andel Institute [copyright holder]
}

}
\keyword{internal}
