% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_energetics_list}
\alias{get_energetics_list}
\title{Get energetics list}
\usage{
get_energetics_list(seahorse_rates)
}
\arguments{
\item{seahorse_rates}{data.table of the input rates}
}
\value{
a list of groups from the data
}
\description{
Makes list of energetics tables
}
\examples{
seahorse_rates <- read_data(seahorse_rates)
split_data <- get_energetics_list(seahorse_rates)
}
