% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{get_energetics_summary}
\alias{get_energetics_summary}
\title{Get energetics summary}
\usage{
get_energetics_summary(energetics)
}
\arguments{
\item{seahorse_rates}{data.table of the input rates}
}
\value{
a list of groups from the data
}
\description{
Makes table of energetics
}
\examples{
replicate_list <- list.files("inst", pattern = "*.xlsx", full.names = TRUE)
seahorse_rates <- read_seahorse(replicate_list)
partitioned_data <- partition_data(seahorse_rates)
energetics <- get_energetics(partitioned_data)
energetics_summary <- get_energetics_summary(energetics_list)
}
