% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.r
\name{make_bioscope_plot}
\alias{make_bioscope_plot}
\title{Bioenergetic scope plot pipeline}
\usage{
make_bioscope_plot(rep_list, ph, pka, buffer)
}
\arguments{
\item{rep_list}{A list of paths to the excel files. One file per replicate}

\item{ph}{pH value for energetics calculation}

\item{pka}{pKa value for energetics calculation}

\item{buffer}{buffer for energetics calculation}
}
\value{
a ggplot
}
\description{
Create a Bioenergetic scope plot from input seahorse rate Excel files
}
\examples{
rep_list <- list.files("inst", pattern = "*.xlsx", full.names = TRUE)
make_bioscope_plot(rep_list, ph, pka, buffer)
}
