% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioscope_plot.R
\name{bioscope_plot}
\alias{bioscope_plot}
\title{Bioenergetic scope plot}
\usage{
bioscope_plot(energetics_summary)
}
\arguments{
\item{energetics_summary}{A summary table of ATP values}
}
\value{
a ggplot
}
\description{
Create a Bioenergetic scope plot
}
\examples{
replicate_list <- c("rep1.xlsx", "rep2.xlsx", "rep3.xlsx")
seahorse_rates <- read_data(replicate_list) |>
                  partition_data() |>
                  get_energetics(ph, pka, buffer) |>
                  get_energetics_summary()
bioscope_plot(seahorse_rates)
}
