% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceas-package.R
\docType{package}
\name{ceas-package}
\alias{ceas}
\alias{ceas-package}
\title{ceas: Cellular Energetics Analysis Software}
\description{
Analysis and visualization of cellular energetics data from Agilent Seahorse XF96. Cellular energetics is how cells make, use, and distribute units of energy (primarily ATP). Measuring real-time cellular energetics is essential to understanding a tissue or cell’s bioenergetic state and fuel dependencies. The Seahorse machine measures a cell’s or matrix’s oxygen consumption rate (OCR) – a proxy of oxidative phosphorylation – and extracellular acidification rate – a proxy of glycolysis. This package offers flexible and fast analysis and plotting capabilities for such data.
}
\keyword{internal}
