% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceas-package.R
\docType{package}
\name{ceas-package}
\alias{ceas}
\alias{ceas-package}
\title{ceas: Cellular Energetics Analysis Software}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'Analysis and visualization of cellular energetics data from Agilent Seahorse XF96. Cellular energetics is how cells make, use, and distribute units of energy (primarily ATP). Measuring real-time cellular energetics is essential to understanding a tissue or cell’s bioenergetic state and fuel dependencies. The Seahorse machine measures a cell’s or matrix’s oxygen consumption rate (OCR) – a proxy of oxidative phosphorylation – and extracellular acidification rate – a proxy of glycolysis. This package offers flexible and fast analysis and plotting capabilities for such data.'
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jamespeapen.github.io/ceas/}
  \item \url{https://github.com/jamespeapen/ceas/}
  \item Report bugs at \url{https://github.com/jamespeapen/ceas/issues/}
}

}
\author{
\strong{Maintainer}: Rachel House \email{rachel.house@vai.org} (\href{https://orcid.org/0000-0003-2176-0431}{ORCID})

Authors:
\itemize{
  \item James Eapen \email{james.eapen@vai.org} (\href{https://orcid.org/0000-0001-6016-3598}{ORCID})
}

Other contributors:
\itemize{
  \item Hui Shen \email{hui.shen@vai.org} (\href{https://orcid.org/0000-0001-9767-4084}{ORCID}) [funder]
  \item Carrie Graveel \email{carrie.graveel@vai.org} (\href{https://orcid.org/0000-0001-7251-5642}{ORCID}) [funder]
  \item Matthew Steensma \email{matthew.steensma@vai.org} (\href{https://orcid.org/0000-0002-9003-6730}{ORCID}) [funder]
  \item Van Andel Institute [copyright holder]
}

}
\keyword{internal}
