% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{get_energetics}
\alias{get_energetics}
\title{Get energetics}
\usage{
get_energetics(partitioned_data, ph, pka, buffer)
}
\arguments{
\item{partitioned_data}{A list of energetics tables from \code{partition_data} function}

\item{ph}{pH value for energetics calculation}

\item{pka}{pKa value for energetics calculation}

\item{buffer}{buffer for energetics calculation}
}
\value{
a list of groups from the data
}
\description{
Makes table of energetics
}
\examples{
replicate_list <- c("rep1.xlsx", "rep2.xlsx", "rep3.xlsx")
energetics <- read_data(replicate_list) |>
                  partition_data() |>
                  get_energetics(ph, pka, buffer)
}
