% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{get_energetics}
\alias{get_energetics}
\title{Calculate ATP Production from OXPHOS and Glycolysis}
\usage{
get_energetics(partitioned_data, ph, pka, buffer)
}
\arguments{
\item{partitioned_data}{a data.table of organized Seahorse OCR and ECAR
rates based on timepoints from the assay cycle. Returned by \code{partition_data}}

\item{ph}{pH value for energetics calculation (for XF Media, 7.5)}

\item{pka}{pKa value for energetics calculation (for XF Media, 6.063)}

\item{buffer}{buffer for energetics calculation (for XF Media, 0.1 mpH/pmol H+)}
}
\value{
a \code{data.table} of glycolysis and OXPHOS rates
}
\description{
Calculates ATP production from glycolysis and OXPHOS at points defined in patitioned_data
}
\examples{
rep_list <- system.file("extdata", package = "ceas") |> list.files(pattern = "*.xlsx", full.names = TRUE)
seahorse_rates <- read_data(rep_list, sheet = 2)
partitioned_data <- partition_data(seahorse_rates)
energetics_list <- get_energetics(partitioned_data, ph = 7.4, pka = 6.093, buffer = 0.1)
}
