% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{get_energetics}
\alias{get_energetics}
\title{Get energetics}
\usage{
get_energetics(seahorse_rates, ph, pka, buffer)
}
\arguments{
\item{seahorse_rates}{data.table of the input rates}
}
\value{
a list of groups from the data
}
\description{
Makes table of energetics
}
\examples{
seahorse_rates <- read_data(seahorse_rates)
energetics <- get_energetics_list(seahorse_rates)
split_data <- get_energetics(seahorse_rates)
}
