% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_energetics_summary}
\alias{get_energetics_summary}
\title{Get energetics summary}
\usage{
get_energetics_summary(energetics)
}
\arguments{
\item{seahorse_rates}{data.table of the input rates}
}
\value{
a list of groups from the data
}
\description{
Makes table of energetics
}
\examples{
seahorse_rates <- read_data(seahorse_rates)
energetics_list <- get_energetics_list(seahorse_rates)
energetics <- get_energetics(seahorse_rates)
energetics_summary <- get_energetics_summary(energetics_list)
}
