% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{partition_data}
\alias{partition_data}
\title{Organize Seahorse Data}
\usage{
partition_data(seahorse_rates)
}
\arguments{
\item{seahorse_rates}{data.table Seahorse OCR and ECAR rates (imported using rep_list function)}

\item{partition_data}{organizes Seahorse OCR and ECAR rates based on defined points during the experiment (e.g. basal OCR before drug addition) - togglable in function if you are modifying Mito Stress Test and Glyco Stress Test measurement numbers (i.e. from 3 measurements to X measurements)}
}
\value{
a list of groups from the data
}
\description{
Makes list of energetics tables
}
\examples{
rep_list <- list.files("result_dir", pattern = "*.xlsx", full.names=TRUE)
seahorse_rates <- read_data(rep_list, sheet=2)
partitioned_data <- partition_data(seahorse_rates)
}
