% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energetics.R
\name{partition_data}
\alias{partition_data}
\title{Get energetics list}
\usage{
partition_data(seahorse_rates)
}
\arguments{
\item{seahorse_rates}{data.table of the input rates}
}
\value{
a list of groups from the data
}
\description{
Makes list of energetics tables
}
\examples{
replicate_list <- c("rep1.xlsx", "rep2.xlsx", "rep3.xlsx")
seahorse_rates <- read_data(replicate_list)
partitioned_data <- partition_data(seahorse_rates)
}
