% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_data}
\alias{read_data}
\title{Input data TODO}
\usage{
read_data(input_data, sheet = 2)
}
\arguments{
\item{input_data}{The name of the sample file}

\item{sheet}{The number of the sheet containing the data. Default is 2
because the output from Seahorse Wave is on sheet 2}
}
\value{
a seahorse_rates table
}
\description{
Reads input data
}
\examples{
seahorse_rates <- read_data("sample1")
}
